moved functions over to utils files
