+++ src/utils/utils.ts
@@ -4,6 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { promises as fs } from 'fs';
+import * as vscode from "vscode";
 
 export async function exists(path: string): Promise<boolean> {
 	try {
@@ -13,5 +14,15 @@ export async function exists(path: string): Promise<boolean> {
 		return false;
 	}
 }
+
+export async function fileExists(uri: vscode.Uri, filename: string): Promise<boolean> {
+    const path = vscode.Uri.joinPath(uri, filename);
+    try {
+		await vscode.workspace.fs.stat(path);
+		return true;
+    } catch {
+      	return false;
+    }
+  }
 
 export class CancelError extends Error { }
